refactor: Separate settings button into reusable widget (#5)
